Moved test code to the standard place.
