+++ src/test/java/com/bitsofproof/supernode/core/ScriptTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.bitsofproof.supernode.test.unit;
+package com.bitsofproof.supernode.core;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -21,11 +21,6 @@ import static org.junit.Assert.assertTrue;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-import com.bitsofproof.supernode.core.ByteUtils;
-import com.bitsofproof.supernode.core.Hash;
-import com.bitsofproof.supernode.core.Script;
-import com.bitsofproof.supernode.core.ValidationException;
-import com.bitsofproof.supernode.core.WireFormat;
 import com.bitsofproof.supernode.model.Tx;
 
 public class ScriptTest
